requirements.txt
asgiref==3.2.3
astroid==2.3.1
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.5
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Django==3.0.2
djangorestframework==3.11.0
Flask==1.1.1
Flask-PyMongo==2.3.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.1
Pillow==7.0.0
pluggy==0.13.0
py==1.8.0
pycodestyle==2.5.0
pylint==2.4.4
pymongo==3.9.0
pyparsing==2.4.2
pytest==5.1.2
python-dotenv==0.12.0
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
virtualenv==16.7.4
wcwidth==0.1.7
Werkzeug==0.15.6
whitenoise==5.0.1
wrapt==1.11.2
zipp==0.6.0
